Strip out external tests
