--- tests/_rfc5_transforms/data_rfc5/from_specification/coordinate_systems/arrayCoordSys.json
@@ -1,10 +1,0 @@
-{
-  "coordinateSystem": {
-    "name": "myDataArray",
-    "axes": [
-      { "name": "k", "type": "array" },
-      { "name": "j", "type": "array" },
-      { "name": "i", "type": "array" }
-    ]
-  }
-}
