--- tests/_rfc5_transforms/data_rfc5/image_example.json
@@ -1,83 +1,0 @@
-{
-  "ome": {
-    "version": "0.5",
-    "multiscales": [
-      {
-        "name": "example",
-        "type": "gaussian",
-        "coordinateSystems": [
-          {
-            "name": "example",
-            "axes": [
-              { "name": "t", "type": "time", "unit": "millisecond" },
-              { "name": "c", "type": "channel" },
-              { "name": "z", "type": "space", "unit": "micrometer" },
-              { "name": "y", "type": "space", "unit": "micrometer" },
-              { "name": "x", "type": "space", "unit": "micrometer" }
-            ]
-          },
-          {
-            "name": "example2",
-            "axes": [
-              { "name": "t", "type": "time", "unit": "millisecond" },
-              { "name": "c", "type": "channel" },
-              { "name": "z", "type": "space", "unit": "micrometer" },
-              { "name": "y", "type": "space", "unit": "micrometer" },
-              { "name": "x", "type": "space", "unit": "micrometer" }
-            ]
-          }
-        ],
-        "datasets": [
-          {
-            "path": "0",
-            "coordinateTransformations": [
-              {
-                "type": "scale",
-                "scale": [1.0, 1.0, 0.5, 0.5, 0.5],
-                "input": "/0",
-                "output": "example"
-              }
-            ]
-          },
-          {
-            "path": "1",
-            "coordinateTransformations": [
-              {
-                "type": "scale",
-                "scale": [1.0, 1.0, 1.0, 1.0, 1.0],
-                "input": "/1",
-                "output": "example"
-              }
-            ]
-          },
-          {
-            "path": "2",
-            "coordinateTransformations": [
-              {
-                "type": "scale",
-                "scale": [1.0, 1.0, 2.0, 2.0, 2.0],
-                "input": "/2",
-                "output": "example"
-              }
-            ]
-          }
-        ],
-        "coordinateTransformations": [
-          {
-            "type": "scale",
-            "scale": [0.1, 1.0, 1.0, 1.0, 1.0],
-            "input": "example",
-            "output": "example2"
-          }
-        ],
-        "metadata": {
-          "description": "the fields in metadata depend on the downscaling implementation. Here, the parameters passed to the skimage function are given",
-          "method": "skimage.transform.pyramid_gaussian",
-          "version": "0.16.1",
-          "args": "[true]",
-          "kwargs": { "multichannel": true }
-        }
-      }
-    ]
-  }
-}
