--- tests/_rfc5_transforms/test_data_rfc5/ngff-rfc5-coordinate-transformation-examples/test_3d_basic.py
@@ -1,35 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_identity() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_scale() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_scale_multiscale() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_sequenceScaleTranslation() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_sequenceScaleTranslation_multiscale() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_translation() -> None:
-    pass
