--- tests/_rfc5_transforms/test_data_rfc5/ngff-rfc5-coordinate-transformation-examples/test_3d_basic_binary.py
@@ -1,15 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_scaleParams() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_translationParams() -> None:
-    pass
