--- tests/_rfc5_transforms/test_data_rfc5/ngff-rfc5-coordinate-transformation-examples/test_3d_simple.py
@@ -1,30 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_affine() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_affine_multiscale() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_affineParams() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_rotation() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_rotationParams() -> None:
-    pass
