--- tests/_rfc5_transforms/test_data_rfc5/ngff-rfc5-coordinate-transformation-examples/test_user_stories.py
@@ -1,25 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_image_registration_3d() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_lens_correction() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_stitched_tiles_2d() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_stitched_tiles_3d() -> None:
-    pass
