--- tests/_rfc5_transforms/test_data_rfc5/ngff/test_coordSystems.py
@@ -1,10 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_arrayCoordSys() -> None:
-    pass
