--- tests/_rfc5_transforms/test_data_rfc5/ngff/test_multiscales_strict.py
@@ -1,20 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_multiscales_example() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_multiscales_example_relative() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_multiscales_transformations() -> None:
-    pass
