--- tests/_rfc5_transforms/test_data_rfc5/ngff/test_subspace.py
@@ -1,15 +1,0 @@
-import pytest
-
-from tests._rfc5_transforms.conftest import get_data_folder
-
-FOLDER = get_data_folder(__file__)
-
-
-@pytest.mark.skip
-def test_subspaceMultidim() -> None:
-    pass
-
-
-@pytest.mark.skip
-def test_subspacePermute() -> None:
-    pass
